requirements.txt
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
SQLAlchemy>=2.0.0
streamlit>=1.30.0
python-dotenv>=1.0.1
pydantic>=2.6.0
